# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem ujian online (bank soal, jadwal, pengerjaan ujian per section, penilaian).",
        "title": "Ujian Backend API",
        "contact": {
            "name": "Tim Backend Ujian"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/soal": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soal"
                ],
                "summary": "List semua soal",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Jumlah per halaman",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all soal successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_soal_dto.SoalListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil data",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Dikirim sebagai multipart/form-data karena mendukung upload gambar untuk soal maupun tiap opsi jawaban.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soal"
                ],
                "summary": "Buat soal baru",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID bank soal (uuid)",
                        "name": "id_bank_soal",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Nomor urut soal dalam bank soal",
                        "name": "no_soal",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Teks soal (bisa HTML)",
                        "name": "soal",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Teks opsi A",
                        "name": "opsi_a",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Teks opsi B",
                        "name": "opsi_b",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Teks opsi C",
                        "name": "opsi_c",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Teks opsi D",
                        "name": "opsi_d",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Teks opsi E",
                        "name": "opsi_e",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Kunci jawaban benar, 1 huruf A-E",
                        "name": "kunci",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Gambar untuk soal",
                        "name": "gambar_soal",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Gambar untuk opsi A",
                        "name": "gambar_a",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Gambar untuk opsi B",
                        "name": "gambar_b",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Gambar untuk opsi C",
                        "name": "gambar_c",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Gambar untuk opsi D",
                        "name": "gambar_d",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Gambar untuk opsi E",
                        "name": "gambar_e",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Create soal successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_soal_dto.SoalResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validasi gagal",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/soal/bank/{bank_soal_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soal"
                ],
                "summary": "List soal dalam satu bank soal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID bank soal (uuid)",
                        "name": "bank_soal_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Jumlah per halaman",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get soal by bank successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_soal_dto.SoalListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil data",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/soal/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload file .xls/.xlsx (maks 10MB) berisi banyak soal sekaligus untuk satu bank soal.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soal"
                ],
                "summary": "Import soal dari file Excel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID bank soal (uuid) tujuan import",
                        "name": "id_bank_soal",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File Excel (.xls/.xlsx)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Import soal berhasil",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_soal_dto.ImportSoalResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "File tidak valid atau import gagal",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/soal/template": {
            "get": {
                "description": "Menghasilkan file .xlsx kosong (hanya header) dengan urutan kolom yang sama persis dengan yang dibaca endpoint import (POST /soal/import) — no_soal, soal, opsi_a-e, kunci, lalu gambar_soal \u0026 gambar_a-e (kolom I dan K sengaja dikosongkan, jangan dihapus/disisipi supaya kolom lain tidak bergeser).",
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "Soal"
                ],
                "summary": "Download template Excel untuk import soal",
                "responses": {
                    "200": {
                        "description": "File template_import_soal.xlsx",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "500": {
                        "description": "Gagal membuat file template",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
...
        "/soal/{id}/restore": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soal"
                ],
                "summary": "Restore soal yang sudah dihapus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID soal (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restore soal successfully",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    },
                    "400": {
                        "description": "Gagal restore",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "backend_internal_helpers.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "backend_internal_modules_soal_dto.ImportSoalErrorDetail": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "row": {
                    "type": "integer"
                }
            }
        },
        "backend_internal_modules_soal_dto.ImportSoalResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_modules_soal_dto.ImportSoalErrorDetail"
                    }
                },
                "import_id": {
                    "type": "string"
                },
                "summary": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "timestamp": {
                    "type": "string"
                },
                "total_failed": {
                    "type": "integer"
                },
                "total_processed": {
                    "type": "integer"
                },
                "total_success": {
                    "type": "integer"
                }
            }
        },
        "backend_internal_modules_soal_dto.SoalListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_modules_soal_dto.SoalResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "backend_internal_modules_soal_dto.SoalResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "gambar_a": {
                    "type": "string"
                },
                "gambar_b": {
                    "type": "string"
                },
                "gambar_c": {
                    "type": "string"
                },
                "gambar_d": {
                    "type": "string"
                },
                "gambar_e": {
                    "type": "string"
                },
                "gambar_soal": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "id_bank_soal": {
                    "type": "string"
                },
                "kunci": {
                    "type": "string"
                },
                "no_soal": {
                    "type": "integer"
                },
                "opsi_a": {
                    "type": "string"
                },
                "opsi_b": {
                    "type": "string"
                },
                "opsi_c": {
                    "type": "string"
                },
                "opsi_d": {
                    "type": "string"
                },
                "opsi_e": {
                    "type": "string"
                },
                "soal": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token JWT dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}